+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 * Adds core implementation
 * Adds regexp paths
 * Adds README
+
+## [0.0.1+1]
+
+* Splits NavigationRouteHandler with a mixin
+* Splits PathRouteHandler with a mixin
+* Removes UnmatchedPathException